+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Master
 
+## Bug Fixes
+
 - Adds support for `csv` and `multi` parameter collectionFormat for query
   parameters.
+
+- Parameters which define both an example (`x-example`) and `items` schema will
+  now use the `x-example` value as the example.
 
 # 0.12.0-beta.2